spi drivers: removed bad cast

+++ arch/efm32/drivers/usart/usart_spi.c
@@ -208,7 +208,7 @@ static bool_t efm32_usart_spi_transfer_tx(struct device_s *dev)
           break;
         }
 
-      cpu_mem_write_32(pv->addr + EFM32_USART_TXDATA_ADDR, endian_le32((uint8_t)word));
+      cpu_mem_write_32(pv->addr + EFM32_USART_TXDATA_ADDR, endian_le32(word));
 
       tr->out = (const void*)((const uint8_t*)tr->out + tr->out_width);
       tr->count--;
